+++ CHANGELOG.md
@@ -9,4 +9,5 @@ Please see each crate's change log below:
 - [`cargo-each`](./crates/cargo-each/CHANGELOG.md)
 - [`cargo-ensure-no-cyclic-deps`](./crates/cargo_ensure_no_cyclic_deps/CHANGELOG.md)
 - [`cargo-ensure-no-default-features`](./crates/cargo-ensure-no-default-features/CHANGELOG.md)
+- [`cargo-ensure-no-unused-workspace-deps`](./crates/cargo-ensure-no-unused-workspace-deps/CHANGELOG.md)
 - [`cargo-heather`](./crates/cargo-heather/CHANGELOG.md)